Cleanup scope related code, add `getSessionFromVSCode` helper
